--- a/justext/core.py
+++ b/justext/core.py
@@ -34,6 +34,7 @@ NO_HEADINGS_DEFAULT = False
 # Short and near-good headings within MAX_HEADING_DISTANCE characters before
 # a good paragraph are classified as good unless --no-headings is specified.
 MAX_HEADING_DISTANCE_DEFAULT = 200
+MAX_PARAGRAPH_LENGTH_DEFAULT = 100000
 PARAGRAPH_TAGS = frozenset({
     'body', 'blockquote', 'caption', 'center', 'col', 'colgroup', 'dd',
     'div', 'dl', 'dt', 'fieldset', 'form', 'legend', 'optgroup', 'option',
@@ -44,6 +45,7 @@ DEFAULT_ENCODING = 'utf8'
 DEFAULT_ENC_ERRORS = 'replace'
 CHARSET_META_TAG_PATTERN = re.compile(br"""<meta[^>]+charset=["']?([^'"/>\s]+)""", re.IGNORECASE)
 GOOD_OR_BAD = {'good', 'bad'}
+GOOD_BAD_NEARGOOD = {'good', 'bad', 'neargood'}
 
 
 class JustextError(Exception):
@@ -310,10 +312,17 @@ def revise_paragraph_classification(paragraphs, max_heading_distance=MAX_HEADING
     has already been called.
     """
 
+    # Attention: This is a fix for the bug in the original code.
+    # Copy the context free class to the class_style
+    # This handles the headings as described in the
+    # documentation
+    for paragraph in paragraphs:
+        paragraph.class_type = paragraph.cf_class
+
     # good headings
     for i, paragraph in enumerate(paragraphs):
         # copy classes
-        paragraph.class_type = paragraph.cf_class
+        # paragraph.class_type = paragraph.cf_class
         if not (paragraph.heading and paragraph.class_type == 'short'):
             continue
         j = i + 1
@@ -371,12 +380,177 @@ def revise_paragraph_classification(paragraphs, max_heading_distance=MAX_HEADING
             j += 1
 
 
+
+def revise_paragraph_classification_fast(paragraphs, max_heading_distance=MAX_HEADING_DISTANCE_DEFAULT):
+    """
+    Optimized context-sensitive paragraph classification. Assumes that classify_pragraphs has already been called.
+    Complexity is O(n), avoiding repeated traversals by pre-computing neighbor information.
+    """
+    n = len(paragraphs)
+    
+    # Attention: This is a fix for the bug in the original code.
+    # Copy the context free class to the class_style
+    # This handles the headings as described in the
+    # documentation
+    for paragraph in paragraphs:
+        paragraph.class_type = paragraph.cf_class
+    
+    # Pre-compute the position of the next good/bad element
+    next_good_or_bad = [n] * n  # Default to end of paragraph list
+    next_good_or_bad_or_neargood = [n] * n
+    
+    # Pre-compute the position of the previous good/bad element
+    prev_good_or_bad = [-1] * n  # Default to beginning of paragraph list
+    prev_good_or_bad_or_neargood = [-1] * n
+    
+    # Step 1: Process good headings
+    # Pre-compute text length for each paragraph
+    text_lengths = [len(p.text) for p in paragraphs]
+    
+    # Pre-compute the position of good paragraphs after each position
+    next_good_pos = [n] * n
+    for i in range(n-1, -1, -1):
+        if paragraphs[i].class_type == 'good':
+            next_good_pos[i] = i
+        elif i < n-1:
+            next_good_pos[i] = next_good_pos[i+1]
+    
+    for i, paragraph in enumerate(paragraphs):
+        if not (paragraph.heading and paragraph.class_type == 'short'):
+            continue
+        
+        # Use pre-computed next_good_pos to quickly find the next good paragraph
+        j = i + 1
+        if j < n and next_good_pos[j] < n:
+            # Calculate distance
+            distance = sum(text_lengths[k] for k in range(i+1, next_good_pos[j]))
+            if distance <= max_heading_distance:
+                paragraph.class_type = 'neargood'
+    
+    # Fill previous element indices from back to front
+    for i in range(n-2, -1, -1):
+        # good or bad
+        if paragraphs[i+1].class_type in GOOD_OR_BAD:
+            prev_good_or_bad[i] = i+1
+            prev_good_or_bad_or_neargood[i] = i+1
+        else:
+            prev_good_or_bad[i] = prev_good_or_bad[i+1]
+            
+        if paragraphs[i+1].class_type in GOOD_BAD_NEARGOOD:
+            prev_good_or_bad_or_neargood[i] = i+1
+        else:
+            prev_good_or_bad_or_neargood[i] = prev_good_or_bad_or_neargood[i+1]
+    
+    # Fill next element indices from front to back
+    for i in range(1, n):
+        # good or bad
+        if paragraphs[i-1].class_type in GOOD_OR_BAD:
+            next_good_or_bad[i] = i-1
+            next_good_or_bad_or_neargood[i] = i-1
+        else:
+            next_good_or_bad[i] = next_good_or_bad[i-1]
+            
+        if paragraphs[i-1].class_type in GOOD_BAD_NEARGOOD:
+            next_good_or_bad_or_neargood[i] = i-1
+        else:
+            next_good_or_bad_or_neargood[i] = next_good_or_bad_or_neargood[i-1]
+
+    # Step 2: Classify short paragraphs
+    new_classes = {}
+    for i, paragraph in enumerate(paragraphs):
+        if paragraph.class_type != 'short':
+            continue
+        
+        # Use pre-computed indices to get neighbor types
+        prev_idx = next_good_or_bad[i]
+        next_idx = prev_good_or_bad[i]
+        
+        prev_neighbour = paragraphs[prev_idx].class_type if prev_idx >= 0 and prev_idx < n else 'bad'
+        next_neighbour = paragraphs[next_idx].class_type if next_idx >= 0 and next_idx < n else 'bad'
+        
+        if prev_neighbour == 'good' and next_neighbour == 'good':
+            new_classes[i] = 'good'
+        elif prev_neighbour == 'bad' and next_neighbour == 'bad':
+            new_classes[i] = 'bad'
+        else:
+            # Check if there are neargood neighbors
+            prev_neargood_idx = next_good_or_bad_or_neargood[i]
+            next_neargood_idx = prev_good_or_bad_or_neargood[i]
+            
+            prev_with_neargood = paragraphs[prev_neargood_idx].class_type if prev_neargood_idx >= 0 and prev_neargood_idx < n else 'bad'
+            next_with_neargood = paragraphs[next_neargood_idx].class_type if next_neargood_idx >= 0 and next_neargood_idx < n else 'bad'
+            
+            if (prev_neighbour == 'bad' and prev_with_neargood == 'neargood') or \
+               (next_neighbour == 'bad' and next_with_neargood == 'neargood'):
+                new_classes[i] = 'good'
+            else:
+                new_classes[i] = 'bad'
+    
+    # Apply new classifications
+    for i, c in new_classes.items():
+        paragraphs[i].class_type = c
+    
+    # Note: Interesting that this is not needed.
+    # Update indices again, as some paragraphs have changed
+    """
+    for i in range(n-2, -1, -1):
+        if paragraphs[i+1].class_type in GOOD_OR_BAD:
+            prev_good_or_bad[i] = i+1
+        else:
+            prev_good_or_bad[i] = prev_good_or_bad[i+1]
+    
+    for i in range(1, n):
+        if paragraphs[i-1].class_type in GOOD_OR_BAD:
+            next_good_or_bad[i] = i-1
+        else:
+            next_good_or_bad[i] = next_good_or_bad[i-1]
+    """
+            
+
+    # Step 3: Revise neargood paragraphs
+    for i, paragraph in enumerate(paragraphs):
+        if paragraph.class_type != 'neargood':
+            continue
+        
+        prev_idx = next_good_or_bad[i]
+        next_idx = prev_good_or_bad[i]
+        
+        prev_neighbour = paragraphs[prev_idx].class_type if prev_idx >= 0 and prev_idx < n else 'bad'
+        next_neighbour = paragraphs[next_idx].class_type if next_idx >= 0 and next_idx < n else 'bad'
+        
+        if (prev_neighbour, next_neighbour) == ('bad', 'bad'):
+            paragraph.class_type = 'bad'
+        else:
+            paragraph.class_type = 'good'
+    
+    # Step 4: More good headings
+    # Pre-compute the position of the first good paragraph after each position
+    next_good_pos = [n] * n
+    for i in range(n-1, -1, -1):
+        if paragraphs[i].class_type == 'good':
+            next_good_pos[i] = i
+        elif i < n-1:
+            next_good_pos[i] = next_good_pos[i+1]
+    
+    for i, paragraph in enumerate(paragraphs):
+        if not (paragraph.heading and paragraph.class_type == 'bad' and paragraph.cf_class != 'bad'):
+            continue
+
+        # Use pre-computed next_good_pos to quickly find the next good paragraph
+        j = i + 1
+        if j < n and next_good_pos[j] < n:
+            # Calculate distance
+            distance = sum(text_lengths[k] for k in range(i+1, next_good_pos[j]))
+            if distance <= max_heading_distance:
+                paragraph.class_type = 'good'
+
+
 def justext(html_text, stoplist, length_low=LENGTH_LOW_DEFAULT,
         length_high=LENGTH_HIGH_DEFAULT, stopwords_low=STOPWORDS_LOW_DEFAULT,
         stopwords_high=STOPWORDS_HIGH_DEFAULT, max_link_density=MAX_LINK_DENSITY_DEFAULT,
         max_heading_distance=MAX_HEADING_DISTANCE_DEFAULT, no_headings=NO_HEADINGS_DEFAULT,
         encoding=None, default_encoding=DEFAULT_ENCODING,
-        enc_errors=DEFAULT_ENC_ERRORS, preprocessor=preprocessor):
+        enc_errors=DEFAULT_ENC_ERRORS, max_paragraph_length=MAX_PARAGRAPH_LENGTH_DEFAULT, preprocessor=preprocessor):
     """
     Converts an HTML page into a list of classified paragraphs. Each paragraph
     is represented as instance of class ˙˙justext.paragraph.Paragraph˙˙.
@@ -386,8 +560,11 @@ def justext(html_text, stoplist, length_low=LENGTH_LOW_DEFAULT,
 
     paragraphs = ParagraphMaker.make_paragraphs(dom)
 
+    if len(paragraphs) > max_paragraph_length:
+        return []
+
     classify_paragraphs(paragraphs, stoplist, length_low, length_high,
         stopwords_low, stopwords_high, max_link_density, no_headings)
-    revise_paragraph_classification(paragraphs, max_heading_distance)
+    revise_paragraph_classification_fast(paragraphs, max_heading_distance)
 
     return paragraphs