@miso-belica 2025-03-05
Thank you for the changes. Can you write tests for these, please? It seems like a lot of code with quite a big impact. Also, it's 3 different changes so maybe we could split the smaller ones into PRs to merge them faster.